+++ src/pages/policy.tsx
@@ -4,9 +4,11 @@ import Head from 'next/head';
 import { createGAOptout } from '../tracking/ga_optout';
 import { GTM_ID } from '../tracking/gtm_id';
 import { addSiteTitleToSuffix } from '../site_title_inserter';
+import { retrieveTranslation } from '../locales/i18n';
+import { SITE_URL } from '../constants/site_data';
+import { PathList } from '../constants/path_list';
 
 import styles from './policy.module.css';
-import { retrieveTranslation } from '../locales/i18n';
 
 const optout = createGAOptout(GTM_ID);
 
@@ -14,6 +16,7 @@ const PolicyPage = (): JSX.Element => {
   const webSiteTitle = retrieveTranslation('website.title');
   const pageTitle = retrieveTranslation('policy.title');
   const titleInHead = addSiteTitleToSuffix(pageTitle);
+  const pageUrl = `${SITE_URL}${PathList.Policy}`;
 
   const [isEnabledOptout, setIsEnabledOptout] = useState(false);
 
@@ -35,6 +38,9 @@ const PolicyPage = (): JSX.Element => {
     <React.Fragment>
       <Head>
         <title>{titleInHead}</title>
+        <meta property="og:title" content={titleInHead} />
+        <meta property="og:url" content={pageUrl} />
+        <meta property="og:type" content="website" />
       </Head>
       <article className={styles.entry}>
         <h2 className={styles.title}>{pageTitle}</h2>
