+++ src/pages/tags/[tag].tsx
@@ -7,6 +7,8 @@ import { EntryList } from '../../components/EntryList';
 import { addSiteTitleToSuffix } from '../../site_title_inserter';
 import { getEntryListByTag, getTagIdList } from '../../entry/entry_gateway';
 import { retrieveTranslation } from '../../locales/i18n';
+import { SITE_URL } from '../../constants/site_data';
+import { PathList } from '../../constants/path_list';
 
 interface Props {
   filteredEntries: Array<EntryValue>;
@@ -19,12 +21,17 @@ export const TagPage = (props: Props): JSX.Element => {
   const webSiteTitle = retrieveTranslation('website.title');
   const titleInHead = addSiteTitleToSuffix(title);
   const description = retrieveTranslation('tags.description', { tag, webSiteTitle });
+  const pageUrl = `${SITE_URL}${PathList.Tags}/${tag}`;
 
   const e = (
     <>
       <Head>
         <title>{titleInHead}</title>
         <meta name="description" content={description} />
+        <meta property="og:title" content={titleInHead} />
+        <meta property="og:description" content={description} />
+        <meta property="og:url" content={pageUrl} />
+        <meta property="og:type" content="website" />
       </Head>
       <EntryList title={title} entries={filteredEntries} />
     </>
